[style] 이전 popupVC 삭제
